moving native utils to a common location
